+++ ext/opentelemetry-ext-requests/tests/test_requests_integration.py
@@ -12,11 +12,10 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import sys
+from unittest import mock
 
 import httpretty
 import requests
-import urllib3
 
 import opentelemetry.ext.requests
 from opentelemetry import context, propagators, trace
@@ -24,6 +23,7 @@ from opentelemetry.ext.requests import RequestsInstrumentor
 from opentelemetry.sdk import resources
 from opentelemetry.test.mock_httptextformat import MockHTTPTextFormat
 from opentelemetry.test.test_base import TestBase
+from opentelemetry.trace.status import StatusCanonicalCode
 
 
 class TestRequestsIntegration(TestBase):
@@ -92,13 +92,8 @@ class TestRequestsIntegration(TestBase):
 
     def test_invalid_url(self):
         url = "http://[::1/nope"
-        exception_type = requests.exceptions.InvalidURL
-        if sys.version_info[:2] < (3, 5) and tuple(
-            map(int, urllib3.__version__.split(".")[:2])
-        ) < (1, 25):
-            exception_type = ValueError
 
-        with self.assertRaises(exception_type):
+        with self.assertRaises(ValueError):
             requests.post(url)
 
         span_list = self.memory_exporter.get_finished_spans()
@@ -109,6 +104,9 @@ class TestRequestsIntegration(TestBase):
         self.assertEqual(
             span.attributes,
             {"component": "http", "http.method": "POST", "http.url": url},
+        )
+        self.assertEqual(
+            span.status.canonical_code, StatusCanonicalCode.INVALID_ARGUMENT
         )
 
     def test_uninstrument(self):
@@ -229,3 +227,75 @@ class TestRequestsIntegration(TestBase):
         span = span_list[0]
 
         self.assertIs(span.resource, resource)
+
+    @mock.patch("requests.Session.send", side_effect=requests.RequestException)
+    def test_requests_exception_without_response(self, *_, **__):
+
+        with self.assertRaises(requests.RequestException):
+            requests.get(self.URL)
+
+        span_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(span_list), 1)
+        span = span_list[0]
+        self.assertEqual(
+            span.attributes,
+            {"component": "http", "http.method": "GET", "http.url": self.URL},
+        )
+        self.assertEqual(
+            span.status.canonical_code, StatusCanonicalCode.UNKNOWN
+        )
+
+    mocked_response = requests.Response()
+    mocked_response.status_code = 500
+    mocked_response.reason = "Internal Server Error"
+
+    @mock.patch(
+        "requests.Session.send",
+        side_effect=requests.RequestException(response=mocked_response),
+    )
+    def test_requests_exception_with_response(self, *_, **__):
+
+        with self.assertRaises(requests.RequestException):
+            requests.get(self.URL)
+
+        span_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(span_list), 1)
+        span = span_list[0]
+        self.assertEqual(
+            span.attributes,
+            {
+                "component": "http",
+                "http.method": "GET",
+                "http.url": self.URL,
+                "http.status_code": 500,
+                "http.status_text": "Internal Server Error",
+            },
+        )
+        self.assertEqual(
+            span.status.canonical_code, StatusCanonicalCode.INTERNAL
+        )
+
+    @mock.patch("requests.Session.send", side_effect=Exception)
+    def test_requests_basic_exception(self, *_, **__):
+
+        with self.assertRaises(Exception):
+            requests.get(self.URL)
+
+        span_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(span_list), 1)
+        self.assertEqual(
+            span_list[0].status.canonical_code, StatusCanonicalCode.UNKNOWN
+        )
+
+    @mock.patch("requests.Session.send", side_effect=requests.Timeout)
+    def test_requests_timeout_exception(self, *_, **__):
+
+        with self.assertRaises(Exception):
+            requests.get(self.URL)
+
+        span_list = self.memory_exporter.get_finished_spans()
+        self.assertEqual(len(span_list), 1)
+        self.assertEqual(
+            span_list[0].status.canonical_code,
+            StatusCanonicalCode.DEADLINE_EXCEEDED,
+        )