+++ opentelemetry-api/CHANGELOG.md
@@ -2,10 +2,12 @@
 
 ## Unreleased
 
-- Rename Observer to ValueObserver in metrics
+- Move stateful from Meter to MeterProvider
   ([#751](https://github.com/open-telemetry/opentelemetry-python/pull/751))
 - Rename Measure to ValueRecorder in metrics
   ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))
+- Rename Observer to ValueObserver
+  ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
 - Add SumObserver in metrics
   ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))