+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,10 +2,14 @@
 
 ## Unreleased
 
-- Rename Observer to ValueObserver
+- Move stateful & resource from Meter to MeterProvider
   ([#751](https://github.com/open-telemetry/opentelemetry-python/pull/751))
 - Rename Measure to ValueRecorder in metrics
   ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))
+- bugfix: byte type attributes are decoded before adding to attributes dict
+  ([#775](https://github.com/open-telemetry/opentelemetry-python/pull/775))
+- Rename Observer to ValueObserver
+  ([#764](https://github.com/open-telemetry/opentelemetry-python/pull/764))
 - Add SumObserver in metrics
   ([#761](https://github.com/open-telemetry/opentelemetry-python/pull/761))